fixed naming for import token
